fix:add styled code  in seperate component
