+++ src/components/MainSection/MainSectionStyled.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+export const Container = styled.main`
+  
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  padding-top: 40px; 
+`;
